+++ src/Timestepping/rhs.jl
@@ -12,17 +12,23 @@ function update_values_at_nodes!(
         fields::NamedTuple, fs, t::Real, iter::VortexFilamentSolver;
         component = Val(:full),  # compute slow + fast components by default
     )
-    scheduler = DynamicScheduler()  # for threading
+    (; to, quantities) = iter
     # (0) Apply affect_t! function (optional)
-    iter.affect_t!(iter, t)
-    # (1) Compute local tangents (this may be used by the forcing)
-    _update_unit_tangents!(iter.tangents, fs; scheduler)
+    if iter.affect_t! !== default_callback()
+        @timeit to "Affect! (fine-grained)" iter.affect_t!(iter, t)
+    end
+    # (1) Compute local tangents (this may be used by the forcing).
+    # In fact, this is currently only used by NormalFluidForcing or in MinimalEnergy mode.
+    if hasproperty(quantities, :tangents)
+        @timeit to "Compute tangents" _update_unit_tangents!(quantities.tangents, fs; scheduler)
+    end
     # (2) Compute velocities and optionally streamfunction values
-    _update_values_at_nodes!(component, iter.fast_term, fields, fs, t, iter)
+    @timeit to "Compute filament velocity" _update_values_at_nodes!(component, iter.fast_term, fields, fs, t, iter)
     # (3) Modify final velocities if running in non-default mode (optional)
     if iter.mode === MinimalEnergy() && haskey(fields, :velocity)
         # Replace velocities (which will be used for advection) with -s⃗′ × v⃗ₛ
-        _minimal_energy_velocities!(fields.velocity, iter.tangents, iter; scheduler)
+        @assert hasproperty(quantities, :tangents)
+        @timeit to "MinimalEnergy mode" _minimal_energy_velocities!(fields.velocity, quantities.tangents, iter; scheduler)
     end
     fields
 end
@@ -34,12 +40,14 @@ function update_values_at_nodes!(vL::VectorOfVectors, fs, t::Real, iter; kws...)
     vL
 end
 
-function _update_unit_tangents!(tangents_all, fs; scheduler)
-    for n in eachindex(fs, tangents_all)
-        f = fs[n]
-        tangents = tangents_all[n]
-        tforeach(eachindex(f, tangents); scheduler) do i
-            tangents[i] = f[i, UnitTangent()]
+function _update_unit_tangents!(tangents_all, fs)
+    @sync for chunk in FilamentChunkIterator(fs)
+        Threads.@spawn for (n, inds, _) in chunk
+            f = fs[n]
+            tangents = tangents_all[n]
+            for i in inds
+                @inbounds tangents[i] = f[i, UnitTangent()]
+            end
         end
     end
     tangents_all