+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.32.7] - 2025-12-15
+
 ### Changed
 
 - Timestepping: avoid computing tangents when not needed.